refactor: api 모듈 구조 변경
